requirements.txt
vkbottle==4.7.0
python-dotenv==1.0.1
aiohttp==3.13.3
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
apscheduler==3.10.4
requests==2.31.0
gigachat==0.1.36
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0
pytz==2023.3
python-dateutil==2.8.2
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
pydantic==2.12.5
